fix: tuist 의존성 수정
