+++ 24th-App-Team-1-iOS/Core/Extensions/Sources/Reactive+Extensions.swift
@@ -6,21 +6,9 @@
 //
 
 import UIKit
-import Networking
 
 import RxSwift
 import RxCocoa
-
-extension ObservableType where Element == Data {
-    public func decodeMap<T: Decodable>(_ type: T.Type) -> Observable<T> {
-        return map { data in
-            guard let decodedData = try? JSONDecoder().decode(T.self, from: data) else {
-                throw WSNetworkError.default(message: "Decoding Error")
-            }
-            return decodedData
-        }
-    }
-}
 
 extension Reactive where Base: UIViewController {
     public var viewWillAppear: ControlEvent<Bool> {
