+++ 24th-App-Team-1-iOS/Core/Util/Project.swift
@@ -13,7 +13,8 @@ let project = Project.makeProject(
     module: .core(.Util),
     targets: [
         .core(module: .Util, dependencies: [
-            .shared(module: .ThirdPartyLib)
+            .shared(module: .ThirdPartyLib),
+            .shared(module: .DesignSystem)
         ])
     ]
 )
