+++ 24th-App-Team-1-iOS/Shared/DesignSystem/Sources/Components/TabBar/WSTabBarViewController.swift
@@ -6,6 +6,7 @@
 //
 
 import UIKit
+import Extensions
 
 import SnapKit
 import RxSwift
